Extract "FeedStoreSpy" to a separate helper
