+++ FeedLoaderTests/Feed cache/CacheFeedUseCaseTests.swift
@@ -144,42 +144,5 @@ class CacheFeedUseCaseTests: XCTestCase {
         return NSError(domain: "any-error", code: 0, userInfo: nil)
     }
     
-    private class FeedStoreSpy: FeedStore {
-        typealias Completion = (Error?) -> Void
-        
-        enum RecievedMessage: Equatable {
-            case deleteCache
-            case insertCache([LocalFeedImage], Date)
-        }
-        
-        var recievedMessages = [RecievedMessage]()
-        private var deletionCompletions = [Completion]()
-        private var insertionCompletions = [Completion]()
-        
-        func deleteCache(completion: @escaping Completion) {
-            deletionCompletions.append(completion)
-            recievedMessages.append(.deleteCache)
-        }
-        
-        func insertCache(with items: [LocalFeedImage], timestamp: Date, completion: @escaping Completion) {
-            insertionCompletions.append(completion)
-            recievedMessages.append(.insertCache(items, timestamp))
-        }
-        
-        func completeInsertionError(_ error: Error, at index: Int = 0) {
-            insertionCompletions[index](error)
-        }
-        
-        func completeInsertionSucessFully(at index: Int = 0) {
-            insertionCompletions[index](nil)
-        }
-        
-        func completeDeletionError(_ error: Error, at index: Int = 0) {
-            deletionCompletions[index](error)
-        }
-        
-        func completeDeletionSucessFully(at index: Int = 0) {
-            deletionCompletions[index](nil)
-        }
-    }
+    
 }
