+++ FeedLoaderTests/Feed cache/LoadFeedFromCacheUseCaseTests.swift
@@ -25,44 +25,6 @@ class LoadFeedFromCacheUseCaseTests: XCTestCase {
         trackMemoryLeak(instance: sut, file: file, line: line)
         return(store, sut)
     }
-    
-    private class FeedStoreSpy: FeedStore {
-        typealias Completion = (Error?) -> Void
-        
-        enum RecievedMessage: Equatable {
-            case deleteCache
-            case insertCache([LocalFeedImage], Date)
-        }
-        
-        var recievedMessages = [RecievedMessage]()
-        private var deletionCompletions = [Completion]()
-        private var insertionCompletions = [Completion]()
-        
-        func deleteCache(completion: @escaping Completion) {
-            deletionCompletions.append(completion)
-            recievedMessages.append(.deleteCache)
-        }
-        
-        func insertCache(with items: [LocalFeedImage], timestamp: Date, completion: @escaping Completion) {
-            insertionCompletions.append(completion)
-            recievedMessages.append(.insertCache(items, timestamp))
-        }
-        
-        func completeInsertionError(_ error: Error, at index: Int = 0) {
-            insertionCompletions[index](error)
-        }
-        
-        func completeInsertionSucessFully(at index: Int = 0) {
-            insertionCompletions[index](nil)
-        }
-        
-        func completeDeletionError(_ error: Error, at index: Int = 0) {
-            deletionCompletions[index](error)
-        }
-        
-        func completeDeletionSucessFully(at index: Int = 0) {
-            deletionCompletions[index](nil)
-        }
-    }
+
 }
